move types to util
